core/requirements.txt
# Family AI Platform - Core Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# HTTP Client
httpx==0.25.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0

# Authentication
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Family AI Specific
openai-whisper==20231117
numpy==1.25.2

# Optional integrations (will cause failures if system dependencies missing)
# Uncomment after system dependencies are installed
# psycopg2-binary==2.9.9
# redis==5.0.1
# matrix-nio==0.24.0
# sounddevice==0.4.6
# pyaudio==0.2.14